feat: 调整web架构层级
