+++ applications/web/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod error;
+pub mod model;
+pub mod service;
