+++ applications/web/src/model/http_model.rs
