+++ applications/web/src/model/mod.rs
@@ -1,0 +1,1 @@
+pub mod http_model;
